# swagger.json
{"openapi":"3.0.0","paths":{"/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/v1/duvidas":{"post":{"operationId":"DuvidasController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDuvidaDto"}}}},"responses":{"201":{"description":""}},"tags":["Dúvidas"]},"get":{"operationId":"DuvidasController_findAll_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["Dúvidas"]}},"/v1/duvidas/{id}":{"get":{"operationId":"DuvidasController_findOne_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Dúvidas"]},"patch":{"operationId":"DuvidasController_update_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDuvidaDto"}}}},"responses":{"200":{"description":""}},"tags":["Dúvidas"]},"delete":{"operationId":"DuvidasController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Dúvidas"]}},"/v1/respostas":{"post":{"operationId":"RespostasController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRespostaDto"}}}},"responses":{"201":{"description":""}},"tags":["Respostas"]}}},"info":{"title":"Ask-Me API","description":"API de dúvidas e respostas para RapidAPI","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{}},"CreateDuvidaDto":{"type":"object","properties":{}},"UpdateDuvidaDto":{"type":"object","properties":{}},"CreateRespostaDto":{"type":"object","properties":{}}}}}